attempted fix for net core 3.1
promoted AssemblyNameEqualityComparer from a private nested class to an internal class
Added a REFACTOR: comment to the minimum effort I think will address the issues with .Net Core 3.1.102
