+++ source/Landorphan.Ioc.ServiceLocation/ServiceLocation/Internal/AssemblyRegistrarRepository.cs
@@ -199,52 +199,6 @@ namespace Landorphan.Ioc.ServiceLocation.Internal
             return rv;
         }
 
-        [SuppressMessage("Microsoft.?", "IDE0048: add parentheses for clarity")]
-        [SuppressMessage(
-            "SonarLint.CodeSmell",
-            "S1067: Expressions should not be too complex",
-            Justification = "Thanks for sharing, I've wrapped this up in a implementation so it never need be done again (MWP)")]
-        private class AssemblyNameEqualityComparer : IEqualityComparer<AssemblyName>
-        {
-            public bool Equals(AssemblyName x, AssemblyName y)
-            {
-                if (x == null)
-                {
-                    return y == null;
-                }
-
-                if (y == null)
-                {
-                    return false;
-                }
-
-                if (x.Name.Equals(y.Name, StringComparison.Ordinal) &&
-                    x.Version.Equals(y.Version) &&
-                    x.CultureInfo.Equals(y.CultureInfo) &&
-                    (ReferenceEquals(x.KeyPair, y.KeyPair) ||
-                     x.KeyPair.IsNotNull() &&
-                     y.KeyPair.IsNotNull() &&
-                     x.KeyPair.PublicKey.SequenceEqual(y.KeyPair.PublicKey)
-                    ))
-                {
-                    return true;
-                }
-
-                return false;
-            }
-
-            public int GetHashCode(AssemblyName obj)
-            {
-                // ReSharper disable once ConditionIsAlwaysTrueOrFalse
-                if (obj != null)
-                {
-                    return obj.GetHashCode();
-                }
-
-                return 0;
-            }
-        }
-
         [SuppressMessage(
             "SonarLint.CodeSmell",
             "S1210: 'Equals' and the comparison operators should be overridden when implementing 'IComparable'",
