--- test/Landorphan.Ioc.ServiceLocation.Tests/TODO.md
@@ -1,4 +1,0 @@
-﻿CA1305 is disabled in the code analysis ruleset but still generates warnings on this project.
-The ruleset specified in the project file
-The ruleset is "None" included in the project file.
-This is true for all of the test projects, something is causing warnings to be emitted in this test project.
